consumer/requirements.txt
kafka-python-ng==2.2.3
elasticsearch==8.17.0